+++ src/page-controller.js
@@ -1,13 +1,12 @@
 import {default as Button} from './components/button.js';
-import {default as FilmCard} from './components/film-card.js';
 import {default as TopRated} from './components/top-rated.js';
 import {default as FilmContainer} from './components/film-container.js';
 import {default as NoSearch} from './components/no-search-result.js';
+import {default as FilmCard} from './components/film-card.js';
 import {render, unrender, Position} from './utils.js';
 import {totalfilm} from "./data.js";
 import {default as Sort} from "./components/sort";
 import {default as MovieController} from "./movie-controller.js";
-const headerContainer = document.querySelector(`.header`);
 
 class PageController {
   constructor(container, film) {
@@ -15,29 +14,6 @@ class PageController {
     this._film = film;
     this._onChangeView = this._onChangeView.bind(this);
     this._onDataChange = this._onDataChange.bind(this);
-  }
-  static renderCard(countFilm, countFilmStart, filmCardContainer, arrFilm, movieController) {
-    const arrFilmSlice = arrFilm.slice(countFilmStart, countFilm);
-    arrFilmSlice.forEach((item) => render(filmCardContainer, new FilmCard(item).getElement(), Position.BEFOREEND));
-    const blockFilmCard = document.querySelectorAll(`.film-card`);
-    for (let item of blockFilmCard) {
-      item.addEventListener(`click`, movieController.onCardTogglerClick);
-    }
-  }
-  static showMoreFilm(countFilm, countFilmStart, filmCardContainer, arrFilm, movieController) {
-    PageController.renderCard(countFilm, countFilmStart, filmCardContainer, arrFilm, movieController);
-    const btnShowFilm = document.querySelector(`.films-list__show-more`);
-    btnShowFilm.addEventListener(`click`, function () {
-      countFilm = countFilm + 5;
-      countFilmStart = countFilmStart + 5;
-      if (countFilm >= totalfilm) {
-        btnShowFilm.style.display = `none`;
-        countFilm = totalfilm;
-        PageController.renderCard(countFilm, countFilmStart, filmCardContainer, arrFilm, movieController);
-      } else {
-        PageController.renderCard(countFilm, countFilmStart, filmCardContainer, arrFilm, movieController);
-      }
-    });
   }
   _onChangeView() {
     this._subscriptions.forEach((subscription) => subscription());
@@ -47,55 +23,15 @@ class PageController {
     this._cards[this._cards.findIndex((card) => card === oldData)] = newData;
     this._renderFilmLists(this._cards, this._generalFilmsList.getElement().querySelector(`.films-list__container`).childElementCount);
   }
-  // Нажатие на карточку
-  // СОРТИРОВКА
-  sortFilm(filmCardContainer, arrFilm, movieController) {
-    const btnSort = document.querySelectorAll(`.sort__button`);
-    const countFilm = 5;
-    const countFilmStart = 0;
-    const linkAddActive = () => {
-      for (let link of btnSort) {
-        if (link.classList.contains(`sort__button--active`)) {
-          link.classList.remove(`sort__button--active`);
-        }
-      }
-    };
-    for (let item of btnSort) {
-      item.addEventListener(`click`, function (evt) {
-        evt.preventDefault();
-        linkAddActive();
-        item.classList.add(`sort__button--active`);
-        const blockFilmCard = document.querySelectorAll(`.film-card`);
-        for (let film of blockFilmCard) {
-          unrender(film);
-        }
-        if (evt.target.dataset.sort === `default`) {
-          const arrFilmDefault = [];
-          for (let i = 0; i < totalfilm; i++) {
-            arrFilmDefault.push(...arrFilm);
-          }
-          PageController.renderCard(countFilm, countFilmStart, filmCardContainer, arrFilmDefault, movieController);
-        } else if (evt.target.dataset.sort === `rating`) {
-          const arrFilmRating = [...arrFilm].sort((filmFirst, filmSecond) => (parseFloat(filmFirst.ratings) - parseFloat(filmSecond.ratings)));
-          PageController.renderCard(countFilm, countFilmStart, filmCardContainer, arrFilmRating, movieController);
-        } else if (evt.target.dataset.sort === `date`) {
-          const arrFilmRDate = [...arrFilm].sort((filmFirst, filmSecond) => (parseInt(filmFirst.year, 10) - parseInt(filmSecond.year, 10)));
-          PageController.renderCard(countFilm, countFilmStart, filmCardContainer, arrFilmRDate, movieController);
-        }
-      });
-    }
-  }
   init() {
+    const headerContainer = document.querySelector(`.header`);
     render(this._container, new Sort().getElement(), Position.BEFOREEND);
     render(this._container, new FilmContainer().getElement(), Position.BEFOREEND);
     const filmContainer = document.querySelector(`.films`);
     const filmList = filmContainer.querySelector(`.films-list`);
     const filmCardContainer = filmList.querySelector(`.films-list__container`);
     render(filmList, new Button().getElement(), Position.BEFOREEND);
-    const movieController = new MovieController(this._container, this._film);
-
-    let countFilm = 5;
-    let countFilmStart = 0;
+    const movieController = new MovieController(filmCardContainer, this._film, totalfilm);
     for (let j = 0; j < 2; j++) {
       render(filmContainer, new TopRated().getElement(), Position.BEFOREEND);
     }
@@ -107,8 +43,6 @@ class PageController {
         item.textContent = `Most comment`;
       }
     });
-    this.sortFilm(filmCardContainer, this._film, movieController);
-    PageController.showMoreFilm(countFilm, countFilmStart, filmCardContainer, this._film, movieController);
     const filmExtraContainer = document.querySelectorAll(`.films-list--extra .films-list__container`);
     const filmExtra = this._film;
     filmExtraContainer.forEach(function () {
